+++ lib/SILOptimizer/Analysis/RegionAnalysis.cpp
@@ -3267,6 +3267,10 @@ bool BlockPartitionState::recomputeExitFromEntry(
           .maybeGetValue();
     }
 
+    RepresentativeValue getRepresentativeValue(Element element) const {
+      return translator.getValueMap().getRepresentativeValue(element);
+    }
+
     bool isClosureCaptured(Element elt, Operand *op) const {
       auto iter = translator.getValueForId(elt);
       if (!iter)
@@ -3503,6 +3507,11 @@ RegionAnalysisValueMap::getRepresentative(Element trackableValueID) const {
 SILValue
 RegionAnalysisValueMap::maybeGetRepresentative(Element trackableValueID) const {
   return getValueForId(trackableValueID)->getRepresentative().maybeGetValue();
+}
+
+RepresentativeValue
+RegionAnalysisValueMap::getRepresentativeValue(Element trackableValueID) const {
+  return getValueForId(trackableValueID)->getRepresentative();
 }
 
 SILIsolationInfo
